Sparky2: enable PPM receiving
The PPM callback is called from the driver space to set the
RFM22B object. This is a weird approach to keeping things
consistent. Generally we don't like to access UAVOs from the
driver layer, but we already access this UAVO from pios_rfm22b_rcvr.

It might make more sense to make sure the PIOS_RCVR layer drirectly
accesses the most recent values in RFM22B_DEV and the callback
is purely used for visibility of the values and debugging.